requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

streamlit==1.33.0
requests==2.31.0

langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3

chromadb==0.4.24

sentence-transformers==2.6.1
transformers==4.39.3
torch>=2.2.0

pypdf==4.2.0
python-dotenv==1.0.1
tiktoken==0.6.0